+++ opencascade-hs/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to the
 
 ## Unreleased
 
+- Add `isDone` to `OpenCascade.BRepBuilderAPI.MakeShape`
+
 ## 0.6.0.0
 
 - Add `OpenCascade.BRepFiletAPI_MakeChamfer`